Reorganise common test functions
