+++ packages/api-tests/test/setup.js
@@ -1,0 +1,22 @@
+const liveServer = require('live-server');
+const {
+  before,
+  after
+} = require('mocha');
+
+const params = {
+  port: 5000,
+  host: '127.0.0.1',
+  root: 'src',
+  open: false,
+  ignore: '*',
+  logLevel: 2
+};
+
+before(() => {
+  liveServer.start(params);
+});
+
+after(() => {
+  liveServer.shutdown();
+});
