+++ packages/api-tests/test/window.spec.js
@@ -1,8 +1,5 @@
 const assert = require('assert');
-const liveServer = require('live-server');
 const {
-  before,
-  after,
   beforeEach,
   afterEach,
   describe,
@@ -10,33 +7,20 @@ const {
 } = require('mocha');
 const testContainer = process.env.MOCHA_CONTAINER;
 const setup = require(`./${testContainer}-test-setup`);
+const {
+  executeAsyncJavascript,
+  selectWindow,
+  openNewWindow
+} = require('./test-helpers');
 
 let app;
-
-const params = {
-  port: 5000,
-  host: '127.0.0.1',
-  root: 'src',
-  open: false,
-  ignore: '*',
-  logLevel: 0
-};
 
 describe('Window API', function(done) {
   const timeout = 60000;
   this.timeout(timeout);
 
-  before(() => {
-    liveServer.start(params);
-  });
-
-  after(() => {
-    liveServer.shutdown();
-  });
-
   beforeEach(() => {
     app = setup(timeout);
-
     return app.start();
   });
 
@@ -45,23 +29,6 @@ describe('Window API', function(done) {
       return app.stop();
     }
   });
-
-  const executeAsyncJavascript = (client, script, ...args) => {
-    // script is passed a callback as its final argument
-    return client.executeAsync(script, ...args);
-  };
-
-  const openNewWindow = (options) => {
-    /* eslint-disable no-undef, no-new */
-    const script = (options, callback) => {
-      ssf.app.ready().then(() => {
-        new ssf.Window(options);
-        setTimeout(() => callback(), 500);
-      });
-    };
-    /* eslint-enable no-undef, no-new */
-    return executeAsyncJavascript(app.client, script, options);
-  };
 
   const callWindowMethod = (method) => {
     /* eslint-disable no-undef */
@@ -88,7 +55,7 @@ describe('Window API', function(done) {
 
   describe('New Window', function() {
     it('Check window constructor opens a new window', function() {
-      return openNewWindow({url: 'about:blank', name: 'test', show: true, child: true}).then((result) => {
+      return openNewWindow(app.client, {url: 'about:blank', name: 'test', show: true, child: true}).then((result) => {
         return app.client.getWindowCount().then((count) => {
           assert.equal(count, 2);
         });
@@ -108,9 +75,8 @@ describe('Window API', function(done) {
       };
 
       return app.client.isVisible('.visible-check')
-        .then(() => openNewWindow(windowOptions))
-        .then(() => app.client.windowHandles())
-        .then((handles) => app.client.window(handles.value[1]))
+        .then(() => openNewWindow(app.client, windowOptions))
+        .then(() => selectWindow(app.client, 1))
         .then(() => app.client.waitForVisible('.visible-check'))
         .then(() => callWindowMethod('getBounds'))
         .then((result) => assert.equal(result.value.x, xValue));
@@ -129,9 +95,8 @@ describe('Window API', function(done) {
       };
 
       return app.client.isVisible('.visible-check')
-        .then(() => openNewWindow(windowOptions))
-        .then(() => app.client.windowHandles())
-        .then((handles) => app.client.window(handles.value[1]))
+        .then(() => openNewWindow(app.client, windowOptions))
+        .then(() => selectWindow(app.client, 1))
         .then(() => app.client.waitForVisible('.visible-check'))
         .then(() => callWindowMethod('getBounds'))
         .then((result) => assert.equal(result.value.y, yValue));
